# auth/mount: fetch single vault resource on read