--- a/website/docs/r/auth_backend.html.md
+++ b/website/docs/r/auth_backend.html.md
@@ -8,6 +8,7 @@ description: |-
 
 # vault\_auth\_backend
 
+This resource enables a new auth method at the given path.
 
 ## Example Usage