@fairclothjm 2024-02-16
Once https://github.com/hashicorp/vault/pull/25499 is merged and the API is tagged we can pull that in