@pree 2024-03-15
Just encountered an issue after upgrading the vault provider that my data resource for `vault_auth_backend` now errors with `* permission denied`.

I had to extend the policy my policy with `sudo` now for this to work:
```hcl
path "sys/auth*" {
  capabilities = ["read", "list", "sudo"]
}
```

This seems like overkill, especially as it's not [documented in the vault API docs](https://developer.hashicorp.com/vault/api-docs/system/auth#read-auth-method-configuration).
Also, the lack of a breaking change documentation isn't nice either ...