@fairclothjm 2024-03-15
@pree Hello, I am sorry you are having trouble! Thanks for pointing this out. We did mention policy changes in the [Changelog](https://github.com/hashicorp/terraform-provider-vault/releases/tag/v4.0.0) and we have published a [v4.0.0 Upgrade Guide](https://registry.terraform.io/providers/hashicorp/vault/latest/docs/guides/version_4_upgrade) but it looks like we missed this nuance.

I think we may be able get a patch in to remove the sudo requirement, however you will still need to update your policy path to use `sys/mounts/auth/*`.