@pree 2024-03-15
Thanks for the reply @fairclothjm. I would love a `# Breaking Change` header in the Changelog to make this more clear.
Removing the `sudo` requirement would be the right approach imho.