@kkronenb 2024-03-19
I have started to play around with this version and so far, it is a night and day difference in our environment. Significantly less CPU (100%->40%) and roughly 3x faster to run though all of our resources. 

Thank you so much and kudos @fairclothjm